+++ mousepacket.h
@@ -7,8 +7,8 @@ class MousePacket : public BusPacket
 {
 public:
 	MousePacket();
-	int x;
-	int y;
+	uint16_t x;
+	uint16_t y;
 };
 
 #endif // MOUSEPACKET_H
